third_party/blink/web_tests/fast/css-grid-layout/grid-shorthand-get-set-expected.txt

This test checks that the 'grid' shorthand is properly parsed and the longhand properties correctly assigned.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Test getting the longhand values when shorthand is set through CSS.
PASS window.getComputedStyle(gridWithNone, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithNone, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithNone, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithNone, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithNone, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithNone, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithTemplate, '').getPropertyValue('grid-template-columns') is "15px"
PASS window.getComputedStyle(gridWithTemplate, '').getPropertyValue('grid-template-rows') is "10px"
PASS window.getComputedStyle(gridWithTemplate, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithTemplate, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithTemplate, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithTemplate, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridInherit, '').getPropertyValue('grid-template-columns') is "15px"
PASS window.getComputedStyle(gridInherit, '').getPropertyValue('grid-template-rows') is "10px"
PASS window.getComputedStyle(gridInherit, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridInherit, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridInherit, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridInherit, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridNoInherit, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridNoInherit, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridNoInherit, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridNoInherit, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridNoInherit, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridNoInherit, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridColumnsAndAutoFlow, '').getPropertyValue('grid-template-columns') is "10px"
PASS window.getComputedStyle(gridColumnsAndAutoFlow, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridColumnsAndAutoFlow, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridColumnsAndAutoFlow, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridColumnsAndAutoFlow, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridColumnsAndAutoFlow, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowAndAutoRows, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithAutoFlowAndAutoRows, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithAutoFlowAndAutoRows, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithAutoFlowAndAutoRows, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithAutoFlowAndAutoRows, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowAndAutoRows, '').getPropertyValue('grid-auto-rows') is "10px"
PASS window.getComputedStyle(gridWithAutoFlowColumnDense, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithAutoFlowColumnDense, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithAutoFlowColumnDense, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithAutoFlowColumnDense, '').getPropertyValue('grid-auto-flow') is "column dense"
PASS window.getComputedStyle(gridWithAutoFlowColumnDense, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowColumnDense, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowDenseRow, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithAutoFlowDenseRow, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithAutoFlowDenseRow, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithAutoFlowDenseRow, '').getPropertyValue('grid-auto-flow') is "dense"
PASS window.getComputedStyle(gridWithAutoFlowDenseRow, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowDenseRow, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowAndAutoRowsAndColumns, '').getPropertyValue('grid-template-columns') is "20px"
PASS window.getComputedStyle(gridWithAutoFlowAndAutoRowsAndColumns, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithAutoFlowAndAutoRowsAndColumns, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithAutoFlowAndAutoRowsAndColumns, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithAutoFlowAndAutoRowsAndColumns, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowAndAutoRowsAndColumns, '').getPropertyValue('grid-auto-rows') is "10px"
PASS window.getComputedStyle(gridWithAutoFlowAndMultipleAutoTracks, '').getPropertyValue('grid-template-columns') is "100px 50px"
PASS window.getComputedStyle(gridWithAutoFlowAndMultipleAutoTracks, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithAutoFlowAndMultipleAutoTracks, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithAutoFlowAndMultipleAutoTracks, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithAutoFlowAndMultipleAutoTracks, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithAutoFlowAndMultipleAutoTracks, '').getPropertyValue('grid-auto-rows') is "20px 10px"
PASS window.getComputedStyle(gridWithRowsAndAutoFlow, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithRowsAndAutoFlow, '').getPropertyValue('grid-template-rows') is "10px"
PASS window.getComputedStyle(gridWithRowsAndAutoFlow, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithRowsAndAutoFlow, '').getPropertyValue('grid-auto-flow') is "column"
PASS window.getComputedStyle(gridWithRowsAndAutoFlow, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithRowsAndAutoFlow, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithRowsNoneAndAutoFlowAndAutoColumn, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithRowsNoneAndAutoFlowAndAutoColumn, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithRowsNoneAndAutoFlowAndAutoColumn, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithRowsNoneAndAutoFlowAndAutoColumn, '').getPropertyValue('grid-auto-flow') is "column"
PASS window.getComputedStyle(gridWithRowsNoneAndAutoFlowAndAutoColumn, '').getPropertyValue('grid-auto-columns') is "10px"
PASS window.getComputedStyle(gridWithRowsNoneAndAutoFlowAndAutoColumn, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithRowsAndAutoFlowDenseColumn, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithRowsAndAutoFlowDenseColumn, '').getPropertyValue('grid-template-rows') is "10px"
PASS window.getComputedStyle(gridWithRowsAndAutoFlowDenseColumn, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithRowsAndAutoFlowDenseColumn, '').getPropertyValue('grid-auto-flow') is "column dense"
PASS window.getComputedStyle(gridWithRowsAndAutoFlowDenseColumn, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithRowsAndAutoFlowDenseColumn, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithColumnsAndDenseAutoFlow, '').getPropertyValue('grid-template-columns') is "10px"
PASS window.getComputedStyle(gridWithColumnsAndDenseAutoFlow, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithColumnsAndDenseAutoFlow, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithColumnsAndDenseAutoFlow, '').getPropertyValue('grid-auto-flow') is "dense"
PASS window.getComputedStyle(gridWithColumnsAndDenseAutoFlow, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithColumnsAndDenseAutoFlow, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithRowsAndAutoFlowAndAutoColumns, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithRowsAndAutoFlowAndAutoColumns, '').getPropertyValue('grid-template-rows') is "20px"
PASS window.getComputedStyle(gridWithRowsAndAutoFlowAndAutoColumns, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithRowsAndAutoFlowAndAutoColumns, '').getPropertyValue('grid-auto-flow') is "column"
PASS window.getComputedStyle(gridWithRowsAndAutoFlowAndAutoColumns, '').getPropertyValue('grid-auto-columns') is "10px"
PASS window.getComputedStyle(gridWithRowsAndAutoFlowAndAutoColumns, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithRowsAndAutoFlowAndMultipleAutoTracks, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithRowsAndAutoFlowAndMultipleAutoTracks, '').getPropertyValue('grid-template-rows') is "100px 50px"
PASS window.getComputedStyle(gridWithRowsAndAutoFlowAndMultipleAutoTracks, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithRowsAndAutoFlowAndMultipleAutoTracks, '').getPropertyValue('grid-auto-flow') is "column"
PASS window.getComputedStyle(gridWithRowsAndAutoFlowAndMultipleAutoTracks, '').getPropertyValue('grid-auto-columns') is "20px 10px"
PASS window.getComputedStyle(gridWithRowsAndAutoFlowAndMultipleAutoTracks, '').getPropertyValue('grid-auto-rows') is "auto"


Test getting wrong values for 'grid' shorthand through CSS (they should resolve to the default: 'none')
PASS window.getComputedStyle(gridWithMisplacedAutoColumn, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithMisplacedAutoColumn, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithMisplacedAutoColumn, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithMisplacedAutoColumn, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithMisplacedAutoColumn, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithMisplacedAutoColumn, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithMisplacedNone1, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithMisplacedNone1, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithMisplacedNone1, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithMisplacedNone1, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithMisplacedNone1, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithMisplacedNone1, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithMisplacedNone2, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithMisplacedNone2, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithMisplacedNone2, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithMisplacedNone2, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithMisplacedNone2, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithMisplacedNone2, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithMisplacedDense, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithMisplacedDense, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithMisplacedDense, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithMisplacedDense, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithMisplacedDense, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithMisplacedDense, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithDuplicatedDense, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithDuplicatedDense, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithDuplicatedDense, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithDuplicatedDense, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithDuplicatedDense, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithDuplicatedDense, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithOnlyDense, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithOnlyDense, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithOnlyDense, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithOnlyDense, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithOnlyDense, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithOnlyDense, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithoutColumnInfo, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithoutColumnInfo, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithoutColumnInfo, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithoutColumnInfo, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithoutColumnInfo, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithoutColumnInfo, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithTwoAutoFlow, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithTwoAutoFlow, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithTwoAutoFlow, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithTwoAutoFlow, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithTwoAutoFlow, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithTwoAutoFlow, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithImplicitAndNoExplicit, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithImplicitAndNoExplicit, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithImplicitAndNoExplicit, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithImplicitAndNoExplicit, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithImplicitAndNoExplicit, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithImplicitAndNoExplicit, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoFlowColumn, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoFlowColumn, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoFlowColumn, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoFlowColumn, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoFlowColumn, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoFlowColumn, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoFlowRow, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoFlowRow, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoFlowRow, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoFlowRow, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoFlowRow, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoFlowRow, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoColumn, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoColumn, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoColumn, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoColumn, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoColumn, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoColumn, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithExtraIdAfterAutoColumn, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithExtraIdAfterAutoColumn, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithExtraIdAfterAutoColumn, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithExtraIdAfterAutoColumn, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithExtraIdAfterAutoColumn, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithExtraIdAfterAutoColumn, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoRow, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoRow, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoRow, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoRow, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoRow, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithExtraIdBeforeAutoRow, '').getPropertyValue('grid-auto-rows') is "auto"
PASS window.getComputedStyle(gridWithExtraIdAfterAutoRow, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(gridWithExtraIdAfterAutoRow, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(gridWithExtraIdAfterAutoRow, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(gridWithExtraIdAfterAutoRow, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(gridWithExtraIdAfterAutoRow, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(gridWithExtraIdAfterAutoRow, '').getPropertyValue('grid-auto-rows') is "auto"


Test getting and setting 'grid' shorthand through JS
PASS getComputedStyle(element, '').getPropertyValue('grid-template-columns') is "10px"
PASS element.style.gridTemplateColumns is "10px"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-rows') is "20px"
PASS element.style.gridTemplateRows is "20px"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-areas') is "none"
PASS element.style.gridTemplateAreas is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-flow') is "row"
PASS element.style.gridAutoFlow is "row"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is "auto"
PASS element.style.gridAutoColumns is "auto"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is "auto"
PASS element.style.gridAutoRows is "auto"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-columns') is "10px"
PASS element.style.gridTemplateColumns is "10px"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-rows') is "[line] 20px"
PASS element.style.gridTemplateRows is "[line] 20px"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-areas') is "\"a\""
PASS element.style.gridTemplateAreas is "\"a\""
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-flow') is "row"
PASS element.style.gridAutoFlow is "row"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is "auto"
PASS element.style.gridAutoColumns is "auto"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is "auto"
PASS element.style.gridAutoRows is "auto"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-columns') is "none"
PASS element.style.gridTemplateColumns is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-rows') is "none"
PASS element.style.gridTemplateRows is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-areas') is "none"
PASS element.style.gridTemplateAreas is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-flow') is "dense"
PASS element.style.gridAutoFlow is "dense"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is "auto"
PASS element.style.gridAutoColumns is "auto"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is "20px"
PASS element.style.gridAutoRows is "20px"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-columns') is "none"
PASS element.style.gridTemplateColumns is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-rows') is "20px"
PASS element.style.gridTemplateRows is "20px"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-areas') is "none"
PASS element.style.gridTemplateAreas is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-flow') is "column"
PASS element.style.gridAutoFlow is "column"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is "10px"
PASS element.style.gridAutoColumns is "10px"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is "auto"
PASS element.style.gridAutoRows is "auto"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-columns') is "none"
PASS element.style.gridTemplateColumns is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-rows') is "none"
PASS element.style.gridTemplateRows is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-areas') is "none"
PASS element.style.gridTemplateAreas is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-flow') is "column dense"
PASS element.style.gridAutoFlow is "column dense"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is "20px"
PASS element.style.gridAutoColumns is "20px"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is "auto"
PASS element.style.gridAutoRows is "auto"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-columns') is "none"
PASS element.style.gridTemplateColumns is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-rows') is "10px"
PASS element.style.gridTemplateRows is "10px"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-areas') is "none"
PASS element.style.gridTemplateAreas is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-flow') is "column"
PASS element.style.gridAutoFlow is "column"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is "20px"
PASS element.style.gridAutoColumns is "20px"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is "auto"
PASS element.style.gridAutoRows is "auto"


Test the initial value
PASS window.getComputedStyle(element, '').getPropertyValue('grid-template-columns') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('grid-template-rows') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('grid-template-areas') is "none"
PASS window.getComputedStyle(element, '').getPropertyValue('grid-auto-flow') is "row"
PASS window.getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is "auto"
PASS window.getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is "auto"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-columns') is 'none'
PASS getComputedStyle(element, '').getPropertyValue('grid-template-rows') is 'none'
PASS getComputedStyle(element, '').getPropertyValue('grid-template-areas') is 'none'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-flow') is 'row'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is 'auto'
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is 'auto'


Test setting grid-template-columns and grid-template-rows back to 'none' through JS
PASS getComputedStyle(element, '').getPropertyValue('grid-template-columns') is "none"
PASS element.style.gridTemplateColumns is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-rows') is "10px"
PASS element.style.gridTemplateRows is "10px"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-areas') is "none"
PASS element.style.gridTemplateAreas is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-flow') is "column"
PASS element.style.gridAutoFlow is "column"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is "20px"
PASS element.style.gridAutoColumns is "20px"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is "auto"
PASS element.style.gridAutoRows is "auto"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-columns') is "none"
PASS element.style.gridTemplateColumns is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-rows') is "none"
PASS element.style.gridTemplateRows is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-areas') is "none"
PASS element.style.gridTemplateAreas is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-flow') is "row"
PASS element.style.gridAutoFlow is "row"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is "auto"
PASS element.style.gridAutoColumns is "auto"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is "auto"
PASS element.style.gridAutoRows is "auto"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-columns') is "none"
PASS element.style.gridTemplateColumns is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-rows') is "20px"
PASS element.style.gridTemplateRows is "20px"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-areas') is "none"
PASS element.style.gridTemplateAreas is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-flow') is "column"
PASS element.style.gridAutoFlow is "column"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is "10px"
PASS element.style.gridAutoColumns is "10px"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is "auto"
PASS element.style.gridAutoRows is "auto"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-columns') is "none"
PASS element.style.gridTemplateColumns is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-rows') is "none"
PASS element.style.gridTemplateRows is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-template-areas') is "none"
PASS element.style.gridTemplateAreas is "none"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-flow') is "row"
PASS element.style.gridAutoFlow is "row"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-columns') is "auto"
PASS element.style.gridAutoColumns is "auto"
PASS getComputedStyle(element, '').getPropertyValue('grid-auto-rows') is "auto"
PASS element.style.gridAutoRows is "auto"


Test the inherit value does not affect gutter properties (grid-*-gap)
PASS getComputedStyle(anotherElement, '').getPropertyValue('grid-column-gap') is "normal"
PASS getComputedStyle(anotherElement, '').getPropertyValue('grid-row-gap') is "normal"
PASS getComputedStyle(anotherElement, '').getPropertyValue('grid-column-gap') is "normal"
PASS getComputedStyle(anotherElement, '').getPropertyValue('grid-row-gap') is "normal"
PASS successfullyParsed is true

TEST COMPLETE
